// app/src/app/models/pdfvaloracion.ts
export interface PdfValoracion {
    Observaciones_val: string;
    pielIntegra: string;
    observacion_Piel: string;
    percepcion_sensorial: number;
    exposicion_humedad: number;
    actividad: number;
    movilidad: number;
    nutricion: number;
    friccion_cizallamiento: number;
    puntajeTotal_Braden: string;
    observaciones_Braden: string;
    total_puntaje_val: string;
    banarse: number;
    comer: number;
    usar_retrete: number;
    subir_escalera: number;
    vestirse: number;
    ctrl_deposicion: number;
    ctrl_miccion: number;
    caminar: number;
    traslado_silla_cama: number;
    PlandeCuidados: string;
    EducacionPaciente: string;
}